requirements.txt
flask
psycopg2-binary
pymongo
sqlparse
ray
